refactor(bytecode): move Instruction enum into isa/def.rs, re-export at module level
The Instruction enum, its INSTRUCTION_WIDTH, op_name/Display impls, and the
enum tests move from isa/mod.rs into a dedicated isa/def.rs; mod.rs keeps the
module overview + submodule wiring and re-exports Instruction/INSTRUCTION_WIDTH,
so every consumer path (crate::isa::Instruction) is unchanged.
